+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ArduinoJson: change log
 =======================
 
+HEAD
+----
+
+* Fix `9.22337e+18 is outside the range of representable values of type 'long'`
+
 v6.19.4 (2022-04-05)
 -------